chore: rename FaqSection to FaqPage
